# BE/admin/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Mô tả API",
        "contact": {
            "email": "email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký tài khoản",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullName",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "a@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Hanoi"
                                    },
                                    "fullName": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng ký thành công"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/api/verify-email": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Xác thực email (kích hoạt tài khoản)",
                "operationId": "30ad750305b3a9051ad8902b774acaa0",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Mã token xác thực được gửi qua email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Abc123xyz456token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tài khoản đã được kích hoạt thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tài khoản đã được kích hoạt thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token không hợp lệ hoặc đã hết hạn.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token không hợp lệ hoặc đã hết hạn."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng nhập người dùng",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Không có quyền"
                    }
                }
            }
        },
        "/api/example": {
            "get": {
                "summary": "Lấy danh sách example",
                "operationId": "8dcd33e8185b7ca83e05be042cb88d93",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "description": "Nhập token theo định dạng: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}